dsim/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsim"
version = "0.1.0"
description = "GPS route simulation: generate motion samples + IQ stream (gps-sdr-sim), then play GPS IQ with speed/bearing control over serial."
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
serial = ["pyserial"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dsim = ["assets/*"]
